[refactor] extract input element to prevent re-render
